+++ projects/sk-scroll-to-top-btn/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## v16.1.0 - Add minDisplayWidth and maxDisplayWidth inputs
+
+- Remove default hide on small screen size
+- Added minDisplayWidth input
+- Added maxDisplayWidth input
+
 ## v16.0.0 - Sync Angular Version
 
 - Sync Angular Version